Se corrige un error en pedidosPendientes
Se corrige para que al borrar un carrito se reintegre correctamente en caso de tratarse de una cantidad muy grande.
Se corrige para que se genere solo un pedido automático por Producto y no por stock

+++ zapateria/src/main/java/dto/DetalleCarritoDTO.java
@@ -3,10 +3,10 @@ package dto;
 public class DetalleCarritoDTO {
 	
 	int id, idCarrito, idProducto, idStock;
-	int cantidad;
+	double cantidad;
 	double precio;
 
-	public DetalleCarritoDTO(int id, int idCarrito, int idProducto, int idStock , int cantidad, double precio) {
+	public DetalleCarritoDTO(int id, int idCarrito, int idProducto, int idStock , double cantidad, double precio) {
 		super();
 		this.id = id;
 		this.idCarrito = idCarrito;
@@ -32,7 +32,7 @@ public class DetalleCarritoDTO {
 		return idStock;
 	}
 	
-	public int getCantidad() {
+	public double getCantidad() {
 		return this.cantidad;
 	}
